refactor agentic pipeline support modules
