+++ algorithm/chat/components/agentic/__init__.py
@@ -1,0 +1,5 @@
+"""Agentic v2 support components.
+
+Keep pipeline orchestration in ``chat.pipelines.agentic_v2`` and place
+protocol/config/review helpers here.
+"""
